rename test utils
